+++ clodoo/clodoo/export_db_model.py
@@ -23,7 +23,7 @@ except ImportError:
 # import pdb
 
 
-__version__ = "2.0.6"
+__version__ = "2.0.7"
 
 
 CACHE = {}