+++ clodoo/clodoo/migrate_odoo_db.py
@@ -41,7 +41,7 @@ import transodoo
 # import pdb
 
 
-__version__ = "2.0.6"
+__version__ = "2.0.7"
 MAX_DEEP = 20
 SYSTEM_MODEL_ROOT = [
     'base.config.',