+++ wok_code/wok_code/scripts/do_odoo_site.py
@@ -80,7 +80,7 @@ TEMPLATE_HTTPS = """#############################################
 </IfModule>
 """
 
-__version__ = "2.0.10"
+__version__ = "2.0.11"
 
 
 class CreateConfig(object):