+++ wok_code/wok_code/scripts/generate_random_codes.py
@@ -8,7 +8,7 @@ from random import random, randint
 import vatnumber
 
 
-__version__ = "2.0.10"
+__version__ = "2.0.11"
 
 
 def gen_vatnumber(opt_args):