+++ wok_code/wok_code/scripts/makepo_it.py
@@ -11,7 +11,7 @@ import argparse
 
 # import pdb
 
-__version__ = "2.0.10"
+__version__ = "2.0.11"
 
 PO_DEFAULT = """
 # Translation of Odoo Server.