+++ wok_code/wok_code/scripts/odoo_translation.py
@@ -19,7 +19,7 @@ except ImportError:
 
 # from python_plus import unicodes
 
-__version__ = "2.0.10"
+__version__ = "2.0.11"
 
 
 MODULE_SEP = "\ufffa"