+++ wok_code/wok_code/scripts/please_apache.py
@@ -8,7 +8,7 @@ except ImportError:
     import configparser as ConfigParser
 
 
-__version__ = "2.0.10"
+__version__ = "2.0.11"
 
 APACHE_TEMPLATE = """##################################################################
 # Odoo service %(odoo_branch)s